backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
google-generativeai==0.8.4
pyserial==3.5
python-dotenv==1.0.1
